added update policy beneficiary cache

+++ ClientServicing/Main/Models/BeneficiaryDetails/BeneficiaryDetailsData.cs
@@ -8,7 +8,7 @@ namespace ClientServicing.Main.Models.BeneficiaryDetails
 {
     public class BeneficiaryDetailsData
     {
-        public int totalAllocated { get; set; }
+        public double totalAllocated { get; set; }
         public int policyNo { get; set; }
         public string auditToken { get; set; }
         public List<BeneficiaryDetailsItems> beneficiaryDetailsItems { get; set; }
